requirements.txt
blinker==1.9.0
cachetools==6.2.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.1.8
confuse==2.0.1
Flask==3.1.2
idna==3.10
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
python-amazon-sp-api==1.9.47
python-dotenv==1.1.1
PyYAML==6.0.2
requests==2.32.5
urllib3==2.5.0
Werkzeug==3.1.3
zipp==3.23.0
